Removed URI from the parent/child link.
Now dealing with the new local URI attribute in CMDI files.
More importantly during the upload of files, where it is used in the matching process between references and files. But also in the import and export process, as well as in other operations regarding linking and unlinking of nodes.

refs #4014
refs #4015