Changed the order in which the nodes are exported. The unlinked / replaced / deleted nodes are dealt with first, and only then the other nodes. This prevents the addition of unnecessary "_2" suffixes to the new nodes, in case they have the same name as the old ones.
Also made the necessary additions to trigger the update of the URLs of the replaced nodes. The lack of this was causing issues when nodes were replaced. If they happened to have the exact same URL (both the old and new nodes), then the crawler wouldn't add the new node to the database, since the URLs are unique.

fixes #4757
fixes #4792